requirements.txt
flask
flask-cors
flask-apscheduler
supabase
requests
python-dotenv 